fix: address Copilot review – fix circular imports and stale closures
- Extract shared types/styles into DrawingCanvas.shared.ts
- Extract useDrawingCanvas hook into DrawingCanvas.hooks.ts
- DrawingCanvas.tsx becomes a pure re-export entrypoint (no circular refs)
- DrawingCanvas.web.tsx and .native.tsx import from .shared / .hooks (no self-import)
- Both platform files re-export named API so @components/DrawingCanvas is complete
- Use functional state updates in handleTouchMove/End/Fill to prevent stale closures
  dropping points on rapid touch events (Copilot findings #4 + #5)
